Fix Redundancy Check
I noticed the redundancy check on line 21 wasn't actually returning on a textarea where I called .autogrow more than once. Apparently doing $e.data('autogrow-enabled') doesn't add anything to $e's data, because $e.data() was {} after it. Specifying a value and checking for it seemed to do the trick. Thanks for the plugin, it is fantastic!